Break circular dependency between the RolloutEvaluator and EvaluateLogBuilder modules
